Fix event notification of removed events.
If more than on event is ready after epoll_wait, it must be ensured that
ready-fds are not signaled anymore if they were removed by a previous
event callback.

Consider the following situation. You have a socket fd and a timer fd
which is used to check if communication is done in time. Now, both fds
get ready after epoll_wait. In the callback of the socket fd the timer
will be cancelled. Now the timer fd must not be used any longer.